Move pic to src/assets/
