+++ src/assets/icons/close.png
