+++ src/components/Loading.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-import loadingGif from "../utils/loading.gif";
+import loadingGif from "../assets/images/loading.gif";
 
 const LoadingWrapper = styled.div`
   display: flex;
