--- src/pages/chatroom/codepen.html
@@ -1,149 +1,0 @@
-<div class="row">
-  <div class="col-4 py-3 px-0 shadow">
-    <table>
-      <tr>
-        <td>
-          <input type="text" placeholder="search" class="form-control" />
-          <i class="fa fa-search fa-2x"></i>
-        </td>
-      </tr>
-      <tr>
-        <td>
-          <div class="box">
-            <img src="https://picsum.photos/50/50/?random=1" />
-            <div>
-              <p>Maine Coon</p>
-              <p>Lorem ipsum dolor sit ...</p>
-            </div>
-          </div>
-        </td>
-      </tr>
-      <tr>
-        <td>
-          <div class="box">
-            <img src="https://picsum.photos/50/50/?random=6" />
-            <div>
-              <p>Samoyed</p>
-              <p>
-                Lorem ipsum dolor sit ... <span class="notice ms-3">9</span>
-              </p>
-            </div>
-          </div>
-        </td>
-      </tr>
-      <tr>
-        <td>
-          <div class="box">
-            <img src="https://picsum.photos/50/50/?random=5" />
-            <div>
-              <p>Border Collie</p>
-              <p>
-                Lorem ipsum dolor sit ... <span class="notice ms-3">5</span>
-              </p>
-            </div>
-          </div>
-        </td>
-      </tr>
-      <tr>
-        <td>
-          <div class="box">
-            <img src="https://picsum.photos/50/50/?random=4" />
-            <div>
-              <p>park0307</p>
-              <p>Lorem ipsum dolor sit ...</p>
-            </div>
-          </div>
-        </td>
-      </tr>
-      <tr>
-        <td>
-          <div class="box">
-            <img src="https://picsum.photos/50/50/?random=3" />
-            <div>
-              <p>park0309</p>
-              <p>Lorem ipsum dolor sit ...</p>
-            </div>
-          </div>
-        </td>
-      </tr>
-      <tr>
-        <td>
-          <div class="box">
-            <img src="https://picsum.photos/50/50/?random=2" />
-            <div>
-              <p>panda</p>
-              <p>Lorem ipsum dolor sit ...</p>
-            </div>
-          </div>
-        </td>
-      </tr>
-    </table>
-  </div>
-  <div class="col-6 px-0 shadow">
-    <table>
-      <tr>
-        <td>
-          <div class="name text-dark">
-            <h6 class="mb-0">
-              <span>Maine Coon</span><i class="fas fa-volume-down"></i>
-            </h6>
-            <div class="ic">
-              <i class="fa fa-search"></i>
-              <i class="fas fa-phone-alt"></i>
-              <i class="fas fa-bars"></i>
-            </div>
-          </div>
-        </td>
-      </tr>
-      <tr>
-        <td>
-          <div class="chat">
-            <div class="other">
-              <img src="https://picsum.photos/50/50/?random=1" />
-              <div class="dialog">
-                Lorem ipsum dolor, sit amet consectetur adipisicing elit. Error,
-                dolore?
-              </div>
-              <p>12:34</p>
-            </div>
-            <div class="self">
-              <div class="dialog">
-                Lorem ipsum dolor, sit amet consectetur adipisicing elit. Error,
-                dolore?
-              </div>
-              <p>12:34</p>
-              <span class="read">read</span>
-            </div>
-            <div class="other">
-              <img src="https://picsum.photos/50/50/?random=1" />
-              <div class="dialog">
-                Lorem ipsum dolor, sit amet consectetur adipisicing elit. Error,
-                dolore?
-              </div>
-              <p>12:34</p>
-            </div>
-          </div>
-        </td>
-      </tr>
-      <tr>
-        <td>
-          <div class="msg">
-            <input
-              type="text"
-              placeholder="you can type something right here, try it!"
-              class="form-control"
-              onchange="tt(this.value)"
-            />
-            <div class="ic">
-              <i class="fas fa-plus"></i>
-              <i class="fas fa-camera"></i>
-              <i class="far fa-image"></i>
-              <span class="float-end"><i class="far fa-smile"></i></span>
-            </div>
-          </div>
-        </td>
-      </tr>
-    </table>
-  </div>
-</div>
-<script src="table.js"></script>
