+++ src/pages/issue/Issue.tsx
@@ -13,7 +13,7 @@ import styled from "styled-components";
 import { useSelector, useDispatch } from "react-redux";
 import firebaseapi from "../../utils/firebaseapi";
 
-import defaultAvatar from "../../utils/DefaultAvatar.png";
+import defaultAvatar from "../../assets/images/DefaultAvatar.png";
 import {
   GoBackWrapper,
   Button,
